// src/app/model/user_get.ts
    export interface UserGetResponse {
        name(name: any): unknown;
        user_id:         number;
        user_email:      string;
        user_pass:       string;
        user_type:       number;
        user_pictrue:    string;
        user_name:       string;
        user_age:        null;
        user_gender:     null;
        user_preference: null;
    }